+++ packages/asc-ui/src/components/ContextMenu/ContextMenuWrapperStyle.ts
@@ -4,6 +4,7 @@ const ContextMenuWrapperStyle = styled.div`
   display: flex;
   align-items: flex-start;
   flex-direction: column;
+  position: relative;
 
   [aria-hidden='true'] {
     display: none;